Number attributes now support large integers

+++ backend/infrahub/graphql/mutations/attribute.py
@@ -1,4 +1,4 @@
-from graphene import Boolean, Field, InputObjectType, Int, String
+from graphene import BigInt, Boolean, Field, InputObjectType, Int, String
 from graphene.types.generic import GenericScalar
 
 from infrahub.core import registry
@@ -47,12 +47,12 @@ class StringAttributeUpdate(BaseAttributeUpdate):
 
 
 class NumberAttributeCreate(BaseAttributeCreate):
-    value = Int(required=False)
+    value = BigInt(required=False)
     from_pool = Field(GenericPoolInput, required=False)
 
 
 class NumberAttributeUpdate(BaseAttributeUpdate):
-    value = Int(required=False)
+    value = BigInt(required=False)
     from_pool = Field(GenericPoolInput, required=False)
